refactor 책임분리
